Change to use params to store Vendor data in Apiary-Scenario. Creation of Asset Resource and update param asset request method.
